requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
Authlib==1.3.0
autopep8==2.0.4
aws-sam-translator==1.85.0
aws-xray-sdk==2.12.1
boto3==1.34.50
botocore==1.34.50
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
cfn-lint==0.85.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
docker==7.0.0
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
execnet==2.0.2
Faker==23.2.1
fastapi==0.110.0
filelock==3.13.1
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jschema-to-python==1.2.3
jsondiff==2.0.0
jsonpatch==1.33
jsonpickle==3.0.3
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
junit-xml==1.9
lazy-object-proxy==1.10.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moto==4.2.13
mpmath==1.3.0
multipart==0.2.4
mypy==1.8.0
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.8.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==23.2
pathable==0.4.3
pbr==6.0.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
py-partiql-parser==0.5.1
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.2
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.2
pytest-check==2.3.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-i18n==0.3.9
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
responses==0.25.0
rfc3339-validator==0.1.4
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
sarif-om==1.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
sshpubkeys==3.3.1
starlette==0.36.3
sympy==1.12
tblib==3.0.0
toml==0.10.2
tomli==2.0.1
typer==0.9.0
types-PyYAML==6.0.12.12
types-toml==0.10.8.7
typing-inspect==0.9.0
typing_extensions==4.10.0
tzlocal==5.2
ujson==5.9.0
Unidecode==1.3.8
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xmltodict==0.13.0
